Stock drift and volotility are now randomly chosen. This may require some fine tuning. The getValue method for Stock class correctly returns double.

+++ src/Trader.java
@@ -4,12 +4,12 @@ import java.util.Set;
 public class Trader {
 
     public String id;
-    public int money;
+    public double money;
 
     // stock id -> num owned
     public HashMap<String, Integer> ownedStocks;
 
-    public Trader(String id, int startingMoney) {
+    public Trader(String id, double startingMoney) {
         this.id = id;
         this.money = startingMoney;
         this.ownedStocks = new HashMap<String, Integer>();
@@ -17,7 +17,7 @@ public class Trader {
 
     // adds the given quantity of the stock to the trader's owned stock
     // uses withdraw money to make sure trader can afford the stock
-    public void addStock(String stockId, int quantity, int value) {
+    public void addStock(String stockId, int quantity, double value) {
         withdrawMoney(quantity * value);
         if(!ownedStocks.containsKey(stockId)) {
             ownedStocks.put(stockId, 0);
@@ -27,7 +27,7 @@ public class Trader {
 
     // withdraws money from the trader's funds
     // helper method for add stock
-    private void withdrawMoney(int amount) {
+    private void withdrawMoney(double amount) {
         if(amount > money) {
             throw new IllegalArgumentException("Cannot withdraw more money than a trader has.");
         }
@@ -38,7 +38,7 @@ public class Trader {
 
     // removes the given quantitiy of the stock from the trader's owned stock
     // makes sure that the trader actually has that much stock to sell
-    public void removeStock(String stockId, int quantity, int value) {
+    public void removeStock(String stockId, int quantity, double value) {
         if(!ownedStocks.containsKey(stockId) || ownedStocks.get(stockId) < quantity) {
             throw new IllegalArgumentException("Cannot withdraw more of a stock than a trader has.");
         }
@@ -48,7 +48,7 @@ public class Trader {
 
     // adds money to the trader's funds
     // helper method for remove stock
-    public void addMoney(int amount) {
+    public void addMoney(double amount) {
         money += amount;
     }
 }
